refactor: 인증 관련 예외 정리
