+++ src/main/java/com/dnd/niceteam/common/Domain.java
@@ -3,4 +3,5 @@ package com.dnd.niceteam.common;
 public enum Domain {
     COMMON,
     AUTH,
+    MEMBER,
 }
