+++ src/main/java/com/dnd/niceteam/error/exception/ErrorCode.java
@@ -4,8 +4,7 @@ import com.dnd.niceteam.common.Domain;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import static com.dnd.niceteam.common.Domain.COMMON;
-import static com.dnd.niceteam.common.Domain.AUTH;
+import static com.dnd.niceteam.common.Domain.*;
 import static javax.servlet.http.HttpServletResponse.*;
 
 @Getter
@@ -17,10 +16,10 @@ public enum ErrorCode {
     INVALID_INPUT_VALUE(COMMON, 3, SC_BAD_REQUEST, "유효하지 않은 입력 값입니다."),
     INVALID_TYPE_VALUE(COMMON, 4, SC_BAD_REQUEST,  "유효하지 않은 타입입니다."),
 
-    MEMBER_NOT_FOUND(COMMON, 5, SC_NOT_FOUND, "존재하지 않는 회원입니다."),
-
     REFRESH_TOKEN_IS_NULL_ERROR(AUTH, 1, SC_NOT_ACCEPTABLE, "로그아웃 상태입니다."),
-    INVALID_TOKEN(AUTH, 4, SC_BAD_REQUEST, "유효하지 않은 토큰입니다."),
+    INVALID_TOKEN(AUTH, 2, SC_BAD_REQUEST, "유효하지 않은 토큰입니다."),
+
+    MEMBER_NOT_FOUND(MEMBER, 1, SC_NOT_FOUND, "존재하지 않는 회원입니다."),
     ;
 
     private final Domain domain;
