--- src/main/java/com/dnd/niceteam/error/exception/MemberNotFoundException.java
@@ -1,7 +1,0 @@
-package com.dnd.niceteam.error.exception;
-
-public class MemberNotFoundException extends BusinessException{
-    public MemberNotFoundException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
-    }
-}
