--- src/main/java/com/dnd/niceteam/error/exception/RefreshTokenIsNullException.java
@@ -1,7 +1,0 @@
-package com.dnd.niceteam.error.exception;
-
-public class RefreshTokenIsNullException extends BusinessException{
-    public RefreshTokenIsNullException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
-    }
-}
