--- src/main/java/com/dnd/niceteam/error/exception/TokenInvalidException.java
@@ -1,7 +1,0 @@
-package com.dnd.niceteam.error.exception;
-
-public class TokenInvalidException extends BusinessException{
-    public TokenInvalidException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
-    }
-}
