+++ src/main/java/com/dnd/niceteam/security/exception/RefreshTokenIsNullException.java
@@ -1,0 +1,10 @@
+package com.dnd.niceteam.security.exception;
+
+import com.dnd.niceteam.error.exception.BusinessException;
+import com.dnd.niceteam.error.exception.ErrorCode;
+
+public class RefreshTokenIsNullException extends BusinessException {
+    public RefreshTokenIsNullException(ErrorCode errorCode, String message) {
+        super(errorCode, message);
+    }
+}
